% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fill_columns.R
\name{fill_columns}
\alias{fill_columns}
\title{fill columns}
\usage{
fill_columns(x, col.name)
}
\arguments{
\item{x}{a data frame}

\item{col.name}{a column name}
}
\value{
a filled data frame
}
\description{
fill columns
}

