% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/newnode_csv.R
\name{newnode_csv}
\alias{newnode_csv}
\title{newnode_csv}
\usage{
newnode_csv(indat = NULL, names_col = NULL, in_col = NULL,
  out_col = NULL, clusters_col = NULL)
}
\arguments{
\item{indat}{the input data.frame}

\item{names_col}{the name of each edge (the boxes)}

\item{in_col}{the name of the nodes that are inputs to each edge with comma seperated vals.  whitespace will be stripped}

\item{out_col}{the nodes that are outputs of each edge.}

\item{clusters_col}{optional}
}
\value{
dataframe representation of the inputs and outputs
}
\description{
newnode_csv
}

