% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/newnode_org_beamer.R, R/newnode_rmd.R
\name{newnode_rmd}
\alias{newnode_org_beamer}
\alias{newnode_rmd}
\title{newnode_rmd}
\usage{
newnode_org_beamer(indat = NULL, names_col = NULL, in_col = NULL,
  out_col = NULL, desc_col = NULL, clusters_col = NULL, todo_col = NULL,
  nchar_to_snip = 40)

newnode_rmd(indat = NULL, names_col = NULL, in_col = NULL,
  out_col = NULL, desc_col = NULL, clusters_col = NULL, todo_col = NULL,
  nchar_to_snip = 40)
}
\arguments{
\item{indat}{the input data.frame}

\item{names_col}{the name of each edge (the boxes)}

\item{in_col}{the name of the nodes that are inputs to each edge with comma seperated vals.  whitespace will be stripped}

\item{out_col}{the nodes that are outputs of each edge.}

\item{desc_col}{description}

\item{clusters_col}{optional column identifying clusters}

\item{todo_col}{optional column with TODO status (DONE and WONTDO will be white, others are red)}

\item{indat}{the input data.frame}

\item{names_col}{the name of each edge (the boxes)}

\item{in_col}{the name of the nodes that are inputs to each edge with comma seperated vals.  whitespace will be stripped}

\item{out_col}{the nodes that are outputs of each edge.}

\item{desc_col}{description}

\item{clusters_col}{optional column identifying clusters}

\item{todo_col}{optional column with TODO status (DONE and WONTDO will be white, others are red)}
}
\value{
character string object that has the DOT language representatio of the input

character string object that has the DOT language representatio of the input
}
\description{
newnode_rmd

newnode_rmd
}

