% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/variable_names_and_labels.r
\name{variable names and labels}
\alias{variable names and labels}
\alias{variable_names_and_labels}
\title{variable_names_and_labels}
\usage{
variable_names_and_labels(infile, datadict = NULL, insert_labels = FALSE)
}
\arguments{
\item{infile}{the full pathname}

\item{datadict}{a dd object (optional will be created if needed)}

\item{insert_labels}{T/F if the output should summarise the value labels}
}
\description{
variable_names_and_labels
}

